simplify redundant files
